extract `fetch-metdata` to separate crate
