+++ Cargo.toml
@@ -3,6 +3,7 @@ members = [
     "cli",
     "codegen",
     "core",
+    "fetch-metadata",
     "lightclient",
     "testing/substrate-runner",
     "testing/test-runtime",
@@ -142,6 +143,7 @@ polkadot-sdk = { version = "0.7", default-features = false }
 # Subxt workspace crates:
 subxt = { version = "0.37.0", path = "subxt", default-features = false }
 subxt-core = { version = "0.37.0", path = "core", default-features = false }
+subxt-fetchmetadata = { version = "0.37.0", path = "fetch-metadata", default_features = false }
 subxt-macro = { version = "0.37.0", path = "macro" }
 subxt-metadata = { version = "0.37.0", path = "metadata", default-features = false }
 subxt-codegen = { version = "0.37.0", path = "codegen" }
