+++ cli/src/commands/chain_spec/mod.rs
@@ -6,7 +6,7 @@ use clap::Parser as ClapParser;
 #[cfg(feature = "chain-spec-pruning")]
 use serde_json::Value;
 use std::{io::Write, path::PathBuf};
-use subxt_codegen::fetch_metadata::Url;
+use subxt_fetchmetadata::url::Url;
 
 mod fetch;
