+++ codegen/Cargo.toml
@@ -13,9 +13,7 @@ description = "Generate an API for interacting with a substrate node from FRAME 
 
 [features]
 default = []
-fetch-metadata = []
-web = ["jsonrpsee?/async-wasm-client", "jsonrpsee?/client-web-transport", "getrandom/js"]
-jsonrpsee = ["dep:jsonrpsee", "url", "dep:frame-metadata", "tokio"]
+web = ["getrandom/js"]
 
 [dependencies]
 codec = { package = "parity-scale-codec", workspace = true, features = ["derive"] }
@@ -26,12 +24,9 @@ quote = { workspace = true }
 syn = { workspace = true }
 scale-info = { workspace = true }
 subxt-metadata = { workspace = true }
-jsonrpsee = { workspace = true, features = ["ws-client", "http-client"], optional = true }
 hex = { workspace = true, features = ["std"] }
-tokio = { workspace = true, features = ["rt-multi-thread"], optional = true }
 thiserror = { workspace = true }
 scale-typegen = { workspace = true }
-url = { workspace = true, optional = true }
 
 # Included if "web" feature is enabled, to enable its js feature.
 getrandom = { workspace = true, optional = true }
