+++ codegen/src/error.rs
@@ -11,9 +11,6 @@ use scale_typegen::TypegenError;
 #[derive(Debug, thiserror::Error)]
 #[non_exhaustive]
 pub enum CodegenError {
-    /// Cannot fetch the metadata bytes.
-    #[error("Failed to fetch metadata, make sure that you're pointing at a node which is providing substrate-based metadata: {0}")]
-    Fetch(#[from] FetchMetadataError),
     /// Cannot decode the metadata bytes.
     #[error("Could not decode metadata, only V14 and V15 metadata are supported: {0}")]
     Decode(#[from] codec::Error),
@@ -60,6 +57,9 @@ pub enum CodegenError {
     /// Error when generating metadata from Wasm-runtime
     #[error("Failed to generate metadata from wasm file. reason: {0}")]
     Wasm(String),
+    /// Other error.
+    #[error("Other error: {0}")]
+    Other(String),
 }
 
 impl CodegenError {
@@ -80,30 +80,4 @@ impl CodegenError {
         let span = self.get_location();
         syn::Error::new(span, msg).into_compile_error()
     }
-}
-
-/// Error attempting to load metadata.
-#[derive(Debug, thiserror::Error)]
-#[non_exhaustive]
-pub enum FetchMetadataError {
-    /// Error decoding from a hex value.
-    #[error("Cannot decode hex value: {0}")]
-    DecodeError(#[from] hex::FromHexError),
-    /// Some SCALE codec error.
-    #[error("Cannot scale encode/decode value: {0}")]
-    CodecError(#[from] codec::Error),
-    /// JSON-RPC error fetching metadata.
-    #[cfg(feature = "fetch-metadata")]
-    #[error("Request error: {0}")]
-    #[cfg(feature = "jsonrpsee")]
-    RequestError(#[from] jsonrpsee::core::ClientError),
-    /// Failed IO when fetching from a file.
-    #[error("Failed IO for {0}, make sure that you are providing the correct file path for metadata: {1}")]
-    Io(String, std::io::Error),
-    /// URL scheme is not http, https, ws or wss.
-    #[error("'{0}' not supported, supported URI schemes are http, https, ws or wss.")]
-    InvalidScheme(String),
-    /// Some other error.
-    #[error("Other error: {0}")]
-    Other(String),
 }
