+++ macro/Cargo.toml
@@ -16,7 +16,7 @@ description = "Generate types and helpers for interacting with Substrate runtime
 [features]
 web = ["subxt-codegen/web"]
 runtime-path = ["polkadot-sdk"]
-jsonrpsee = ["subxt-codegen/jsonrpsee"]
+runtime-metadata-insecure-url = ["subxt-fetchmetadata/url"]
 
 [lib]
 proc-macro = true
@@ -27,7 +27,8 @@ darling = { workspace = true }
 proc-macro-error2 = { workspace = true }
 syn = { workspace = true }
 quote = { workspace = true }
-subxt-codegen = { workspace = true, features = ["fetch-metadata"] }
+subxt-codegen = { workspace = true }
+subxt-fetchmetadata = { workspace = true }
 scale-typegen = { workspace = true }
 polkadot-sdk = { workspace = true, optional = true, features = ["sp-io", "sc-executor-common", "sp-state-machine", "sp-maybe-compressed-blob", "sc-executor"] }
