+++ macro/src/lib.rs
@@ -15,9 +15,7 @@ use scale_typegen::typegen::{
     settings::substitutes::path_segments,
     validation::{registry_contains_type_path, similar_type_paths_in_registry},
 };
-use subxt_codegen::{
-    fetch_metadata::fetch_metadata_from_file_blocking, CodegenBuilder, CodegenError, Metadata,
-};
+use subxt_codegen::{CodegenBuilder, CodegenError, Metadata};
 use syn::{parse_macro_input, punctuated::Punctuated};
 
 #[cfg(feature = "runtime-path")]
@@ -239,15 +237,13 @@ fn fetch_metadata(args: &RuntimeMetadataArgs) -> Result<subxt_codegen::Metadata,
             let root = std::env::var("CARGO_MANIFEST_DIR").unwrap_or_else(|_| ".".into());
             let root_path = std::path::Path::new(&root);
             let path = root_path.join(rest_of_path);
-            fetch_metadata_from_file_blocking(&path)
+            subxt_fetchmetadata::file::get_blocking(&path)
                 .and_then(|b| subxt_codegen::Metadata::decode(&mut &*b).map_err(Into::into))
-                .map_err(|e| CodegenError::from(e).into_compile_error())?
+                .map_err(|e| CodegenError::Other(e.to_string()).into_compile_error())?
         }
-        #[cfg(feature = "jsonrpsee")]
+        #[cfg(feature = "runtime-metadata-insecure-url")]
         (None, Some(url_string)) => {
-            use subxt_codegen::fetch_metadata::{
-                fetch_metadata_from_url_blocking, MetadataVersion, Url,
-            };
+            use subxt_fetchmetadata::url::{get_blocking, MetadataVersion, Url};
 
             let url = Url::parse(url_string).unwrap_or_else(|_| {
                 abort_call_site!("Cannot download metadata; invalid url: {}", url_string)
@@ -258,16 +254,16 @@ fn fetch_metadata(args: &RuntimeMetadataArgs) -> Result<subxt_codegen::Metadata,
                 false => MetadataVersion::Latest,
             };
 
-            fetch_metadata_from_url_blocking(url, version)
-                .map_err(CodegenError::from)
+            get_blocking(url, version)
+                .map_err(|e| CodegenError::Other(e.to_string()))
                 .and_then(|b| subxt_codegen::Metadata::decode(&mut &*b).map_err(Into::into))
                 .map_err(|e| e.into_compile_error())?
         }
-        #[cfg(not(feature = "jsonrpsee"))]
+        #[cfg(not(feature = "runtime-metadata-insecure-url"))]
         (None, Some(_)) => {
             abort_call_site!(
-                    "The 'runtime_metadata_insecure_url' attribute is not supported without the 'jsonrpsee' feature"
-                )
+                "'runtime_metadata_insecure_url' requires the 'runtime-metadata-insecure-url' feature to be enabled"
+            )
         }
         #[cfg(feature = "runtime-path")]
         (None, None) => {
