+++ macro/src/wasm_loader.rs
@@ -12,15 +12,15 @@ use polkadot_sdk::{
     sp_maybe_compressed_blob::{self, CODE_BLOB_BOMB_LIMIT},
     sp_state_machine,
 };
-use subxt_codegen::{fetch_metadata::fetch_metadata_from_file_blocking, CodegenError, Metadata};
+use subxt_codegen::{CodegenError, Metadata};
 
 /// Result type shorthand
 pub type WasmMetadataResult<A> = Result<A, CodegenError>;
 
 /// Uses wasm artifact produced by compiling the runtime to generate metadata
 pub fn from_wasm_file(wasm_file_path: &Path) -> WasmMetadataResult<Metadata> {
-    let wasm_file = fetch_metadata_from_file_blocking(wasm_file_path)
-        .map_err(Into::<CodegenError>::into)
+    let wasm_file = subxt_fetchmetadata::file::get_blocking(wasm_file_path)
+        .map_err(|e| CodegenError::Other(e.to_string()))
         .and_then(maybe_decompress)?;
     call_and_decode(wasm_file)
 }
